sgp4x: Expose the raw VOC and NOX values as sensors
The SGP4x chips define an algorithm to convert the raw sensor data
to an "index".  The index is quite useful, but, depending on the application,
it might not be the right thing.  Expose the raw data as well.

The ticks are "raw data proportional to the logarithm of the resistance of the
sensing element". Figures 2 and 3 of the datasheet give some guidance for the
how the raw output relates to the actual concentration of gasses in the air.